+++ src/emitters.ts
@@ -18,6 +18,9 @@ export const ccloudOrganizationChanged = new vscode.EventEmitter<void>();
 /** Fired whenever the list of direct connections changes. */
 export const directConnectionsChanged = new vscode.EventEmitter<void>();
 
+/** Fired when websocket event for a CREATED direct connection environment is received. */
+export const directConnectionCreated = new vscode.EventEmitter<ConnectionId>();
+
 export const localKafkaConnected = new vscode.EventEmitter<boolean>();
 export const localSchemaRegistryConnected = new vscode.EventEmitter<boolean>();