+++ src/viewProviders/resources.ts
@@ -14,6 +14,7 @@ import {
   ccloudConnected,
   ccloudOrganizationChanged,
   connectionStable,
+  directConnectionCreated,
   directConnectionsChanged,
   localKafkaConnected,
   localSchemaRegistryConnected,
@@ -95,6 +96,8 @@ export class ResourceViewProvider implements vscode.TreeDataProvider<ResourceVie
    */
   environmentsMap: Map<string, CCloudEnvironment | LocalEnvironment | DirectEnvironment> =
     new Map();
+
+  private possiblyLoadingConnectionIds: Set<ConnectionId> = new Set();
 
   // env id -> preannounced loading state coming from sidecar websocket
   // events to us via connectionUsable emitter.
@@ -208,6 +211,9 @@ export class ResourceViewProvider implements vscode.TreeDataProvider<ResourceVie
         ContainerTreeItem<LocalKafkaCluster | LocalSchemaRegistry>,
         DirectEnvironment[],
       ] = await Promise.all(resourcePromises);
+
+      this.assignIsLoading(directEnvironments);
+
       children = [ccloudContainer, localContainer, ...directEnvironments];
 
       if (this.forceDeepRefresh) {
@@ -291,6 +297,15 @@ export class ResourceViewProvider implements vscode.TreeDataProvider<ResourceVie
 
   /** Set up event listeners for this view provider. */
   setEventListeners(): vscode.Disposable[] {
+    const newlyCreatedConnectionSub: vscode.Disposable = directConnectionCreated.event(
+      (connectionId: ConnectionId) => {
+        logger.debug(
+          "resourcesView: directConnectionCreated event fired, marking as newly created",
+          { connectionId },
+        );
+        this.possiblyLoadingConnectionIds.add(connectionId);
+      },
+    );
     const ccloudConnectedSub: vscode.Disposable = ccloudConnected.event((connected: boolean) => {
       logger.debug("ccloudConnected event fired", { connected });
       // No need to force a deep refresh when the connection status changes because the
@@ -329,6 +344,8 @@ export class ResourceViewProvider implements vscode.TreeDataProvider<ResourceVie
 
     const connectionUsableSub: vscode.Disposable = connectionStable.event((id: ConnectionId) => {
       logger.debug("connectionStable event fired", { id });
+      // if is stable, then is definitely not loading anymore.
+      this.possiblyLoadingConnectionIds.delete(id);
       this.refreshConnection(id, false);
     });
 
@@ -355,6 +372,7 @@ export class ResourceViewProvider implements vscode.TreeDataProvider<ResourceVie
     );
 
     return [
+      newlyCreatedConnectionSub,
       ccloudConnectedSub,
       ccloudOrganizationChangedSub,
       directConnectionsChangedSub,
@@ -474,6 +492,36 @@ export class ResourceViewProvider implements vscode.TreeDataProvider<ResourceVie
     // clear from any previous search filter
     this.searchMatches = new Set();
     this.totalItemCount = 0;
+  }
+
+  /**
+   * Assign the loading state to all direct environments.
+   * Takes note as to which connections should be considered
+   * "loading" from the {@link possiblyLoadingConnectionIds} set, and
+   * possibly reassigns the environment's isLoading accordingly.
+   * */
+  assignIsLoading(directEnvs: DirectEnvironment[]): void {
+    directEnvs.forEach((env) => {
+      const isNewlyCreated = this.possiblyLoadingConnectionIds.has(env.connectionId);
+
+      if (isNewlyCreated) {
+        // Connection is new, so we want to show the spinny icon. Leave env.isLoading as is.
+        // Remove the connection ID from the set so we will clean any isLoading in future repaints.
+        this.possiblyLoadingConnectionIds.delete(env.connectionId);
+
+        logger.debug(`assignIsLoading() leaving isLoading=true for direct environment ${env.id}`, {
+          connectionId: env.connectionId,
+        });
+      } else {
+        // Environment objects w/o clusters are default to isLoading = true, which makes the spinny icon.
+        // Only want the spinning icon for actual new connections which should then get followup
+        // websocket events marking them as no longer loading.
+        env.isLoading = false;
+        logger.debug(`assignIsLoading() setting isLoading=false for direct environment ${env.id}`, {
+          connectionId: env.connectionId,
+        });
+      }
+    });
   }
 }